Test that file exist before delete it

+++ plexus-container-default.old/src/java/org/codehaus/plexus/util/FileUtils.java
@@ -1019,6 +1019,11 @@ public class FileUtils
     public static void forceDelete( final File file )
         throws IOException
     {
+        if ( ! file.exists() )
+        {
+            return;
+        }
+
         if ( file.isDirectory() )
         {
             deleteDirectory( file );
@@ -1041,6 +1046,11 @@ public class FileUtils
     public static void forceDeleteOnExit( final File file )
         throws IOException
     {
+        if ( ! file.exists() )
+        {
+            return;
+        }
+
         if ( file.isDirectory() )
         {
             deleteDirectoryOnExit( file );
